# Fix tractography rendering performance on macOS