--- a/src/gui/mrview/tool/tractography/tractogram.h
+++ b/src/gui/mrview/tool/tractography/tractogram.h
@@ -144,6 +144,9 @@ namespace MR
             vector<vector<GLint> > original_track_sizes;
             vector<vector<GLint> > original_track_starts;
             vector<size_t> num_tracks_per_buffer;
+            // EBOs and indices for chunks of tracks
+            std::vector<GLuint> element_buffers;
+            std::vector<GLsizei> element_counts;
             GLint sample_stride;
             bool vao_dirty;