@bjeurissen 2025-10-04
Would be good to check if the same fix applies to #2379. At least the symptoms look identical, so I would think there is a good chance.