@jdtournier 2025-10-06
:+1: Works fine on Ubuntu 22.04, running on a 12GB NVIDIA GeForce RTX 4070. Get roughly 50 FPS for 1 million streamlines with streamtube lighting in volume render mode.

I'll try it on my home system when I get home, just to check with an AMD GPU.