@jdtournier 2025-10-07
Also works fine on my AMD-based home system 👍